+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+- Crash when using `react-native-reanimated` along with `bitmovin-player-react-native` and playing a live stream
+
 ## [1.0.0] - 2025-08-04
 
 ### Breaking Change