+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Crash when using `react-native-reanimated` along with `bitmovin-player-react-native` and playing a live stream
 - Expo config plugin feature configurations taking no effect
+- Example application crash on tvOS simulator
 
 ## [1.0.0] - 2025-08-04