+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unchecked]
 
+### Changed
+
+- Update Bitmovin's native Android SDK version to `3.123.0`
+- Update Bitmovin's native iOS SDK version to `3.94.1`
+
 ### Fixed
 
 - Crash when using `react-native-reanimated` along with `bitmovin-player-react-native` and playing a live stream