+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unchecked]
+## [1.1.0]
 
 ### Changed